website-redesign-agent/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
crewai>=0.28.0
requests>=2.28.0
beautifulsoup4>=4.12.0
httpx>=0.24.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
anthropic>=0.5.0
langchain>=0.0.267
langchain-community>=0.0.9
langchain-anthropic>=0.0.5
openai>=1.1.1
